fashion_project/requirements.txt
Django
djangorestframework
Pillow
serpapi
fashion-clip
insightface
opencv-python
numpy
torch
streamlit
requests
python-dotenv
onnxruntime
pandas